src/pyproject.toml
[tool.poetry]
name = "django-webapp-starter"
version = "0.0.1"
description = "DESCRIPTION OF YOUR DJANGO APP"
authors = ["AUTHOR <YOURMAIL@DOMAIN.org>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.7.2"
astroid = "^3.0.0"
beautifulsoup4 = "^4.12.2"
celery = "^5.3.5"
certifi = "^2024.8.30"
charset-normalizer = "^3.2.0"
click = "^8.1.7"
colorama = "^0.4.6"
colorlog = "^6.7.0"
cssbeautifier = "^1.14.9"
dill = "^0.3.7"
dj-database-url = "^2.1.0"
django = "^5.1.3"
django-allauth = "^65.2.0"
django-celery-beat = "^2.5.0"
django-cleanup = "^9.0.0"
django-extensions = "^3.2.3"
django-htmx = "^1.16.0"
django-redis = "^5.4.0"
django-split-settings = "^1.2.0"
django-use-email-as-username = "^1.4.0"
editorconfig = "^0.12.3"
gprof2dot = "^2024.6.6"
gunicorn = "^23.0.0"
html-tag-names = "^0.1.2"
html-void-elements = "^0.1.0"
idna = "^3.4"
iniconfig = "^2.0.0"
jsbeautifier = "^1.14.9"
lazy-object-proxy = "^1.9.0"
libsass = "^0.23.0"
mccabe = "^0.7.0"
mypy-extensions = "^1.0.0"
packaging = "^24.2"
pathspec = "^0.12.1"
pillow = "^11.0.0"
platformdirs = "^4.3.6"
pluggy = "^1.3.0"
psycopg2 = "^2.9.7"
pytest = "^8.3.3"
pytest-django = "^4.5.2"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
redis = "^5.0.1"
requests = "^2.31.0"
six = "^1.16.0"
soupsieve = "^2.5"
sqlparse = "^0.5.1"
tomlkit = "^0.13.2"
toposort = "^1.10"
tqdm = "^4.66.1"
tzdata = "^2024.2"
urllib3 = "^2.0.5"
wrapt = "^1.15.0"
more-itertools = "^10.5.0"
typer = "^0.13.0"
djlint = "^1.36.1"
regex = "^2024.11.6"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.3"
django-debug-toolbar = "^4.2.0"
djlint = "^1.34.0"
isort = "^5.12.0"
pep8 = "^1.7.1"
pre-commit = "^4.0.1"
flake8 = "^7.1.1"
pycodestyle = "^2.11.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_list = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
